requirement.txt
numpy==2.2.5
opencv-python==4.11.0.86
matplotlib==3.10.3
numpy-stl==3.2.0
pillow==11.2.1
python-utils==3.9.1
